add Record and Message classes
`record.message` is now an instanceof Message, and implements toString and
toJSON. This improves how records output when using the JSON format, such that
the args aren't printed as the `args` property, and stringified in the
`message` property. Now, if output is %O (json), the `message` property will
be an array of all the args passed. If `%(message)s` is used, it will still
stringify all the args into a log message.